qemu: Rewrite, replace injected init script with ttyS0 login
There does not seem to be an adequate non-root way of injecting an init script
into an overlay: nbd requires root, libguestfs has heavy dependencies and does
not currently work in Ubuntu, and cloud-image is too strong an assumption.

So instead, assume that VMs satisfy one of:

 - They open a root shell on ttyS1

 - They have a serial terminal on ttyS0 where you can log in with specified
   user and password. This will then be used to start a root shell on ttyS1, to
   reduce this to the first case and to not assume that ttyS0 stays
   operational throughout the whole test.

Also add some common functionality to VirtSubproc for opening a FD to a Unix
socket, and expecting some output (within a timeout) on such a socket.

This now also uses argparse instead of optparse, which opens the possibility of
putting command line arguments into files with the @ prefix.