Skip package installation if testbed does not have root
Make installation of packages/dependencies print a warning instead of failing
immediately when root is not available on the testbed. This allows tests to be
run with the "null" runner locally as user with some restrictions.

Check if dpkg-source is available before installing dpkg-dev, to avoid the
warning in that simple case.

Closes: #742242