Fix invocation of tests so that they retain the default SIGINT handler
We previously ran the test in a background bash subshell, so that we could
inject a trap to forward SIGTERM to the child. But background shells set the
signal handlers for INT to SIG_IGN, so that tests can't use that signal any
more.

Avoid that by running the test in the foreground again and dropping the trap
(which was never very reliable as it wouldn't terminate grandchildren). Replace
this with putting the test invocation script into a new process group and
killing the whole group on timeout.

Closes: #761596